requirements.txt
ipykernel==6.30.1
jupyter-client==8.6.3
matplotlib==3.10.6
numpy==2.2.6
openai==2.1.0
opencv-python==4.12.0.88
pandas==2.3.3
pillow==11.3.0
pyyaml==6.0.2
qwen-vl-utils==0.0.14
requests==2.32.4
scikit-image==0.25.2
scikit-learn==1.7.1
scipy==1.16.0
sympy==1.13.3
torch==2.8.0
torchvision==0.23.0
tqdm==4.67.1
trafilatura==2.0.0
